+++ ui/index.html
@@ -89,6 +89,8 @@
                     <ul class="dropdown-menu dropdown-user">
                         <li id="usersettingsbutton"><a href="#usersettings"><i class="fa fa-user fa-fw"></i> User Settings</a>
                         </li>
+                        <li id="accounting_settings_button"><a href="#accounting_settings"><i class="fa fa-book fa-fw"></i> Accounting Settings</a>
+                        </li>
                         <li id="settingsbutton"><a href="#settings"><i class="fa fa-gear fa-fw"></i> Settings</a>
                         </li>
                         <li class="divider"></li>
